# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Marveen Dashboard API
  version: 1.0.0
  description: |
    Internal REST API for the Marveen fleet dashboard and inter-agent communication.
    All endpoints require a Bearer token (`store/.dashboard-token`).

    **Tiers**
    - Tier 1 (full schema): core inter-agent endpoints — memories, messages, kanban,
      agents, skills, approvals, blackboard, daily-log, skill-usage.
    - Tier 2 (medium): schedules, ideas, artifacts, token-usage, connectors, recall, autonomy.
    - Tier 3 (x-internal): status checks, dashboard-internal routes, static assets.
servers:
  - url: http://localhost:3420/api/v1
    description: Canonical API base (v1)
  - url: http://localhost:3420/api
    description: Legacy API base (deprecated, Deprecation + Sunset headers returned)
security:
  - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Token stored in `store/.dashboard-token`
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: >-
            Machine-readable snake_case error token. Canonical values are listed
            in the enum; additional domain-specific tokens may appear in future
            API versions.
          enum:
            - not_found
            - required
            - invalid_value
            - forbidden
            - unauthorized
            - conflict
            - limit_exceeded
            - internal_error
            - parse_error
            - not_supported
            - timeout
            - disabled
            - managed_settings_missing
            - upstream_error
            - sender_not_in_allowlist
            - federation_disabled
            - unknown_query_parameter





        hint:
          type: string
          description: >-
            Optional human-readable debugging note. Present when the server has
            extra context that helps the caller fix the request (e.g. which
            parameter name the filter expects, or why a value was rejected).
            Clients must not rely on its exact text; treat it as informational.
        field:
          type: string
          description: >-
            Name of the request field that caused the validation failure.
            Present on 4xx validation errors when the problem can be attributed
            to a single input field.
      example:
        error: "required"
        field: "agent_id"
    OkResponse:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
...
    AgentMessage:
      type: object
      required: [id, from_agent, to_agent, content, status, created_at]
      properties:
        id:
          type: integer
        from_agent:
          type: string
        to_agent:
          type: string
        content:
          type: string
        status:
          type: string
          enum: [pending, delivered, done, failed]
        result:
          type: [string, "null"]
        origin_note:
          type: [string, "null"]
          maxLength: 120
        created_at:
          type: integer
        delivered_at:
          type: [integer, "null"]
        completed_at:
          type: [integer, "null"]
        trace_id:
          type: [string, "null"]
        span_id:
          type: [string, "null"]
      example:
        id: 3715
        from_agent: "agent-b"
        to_agent: "agent-a"
        content: "feat/zod-config-schema kész -- 23ec40d, 24 teszt zöld"
        status: "pending"
        result: null
        origin_note: null
        created_at: 1787487278
...
paths:
...
  /messages:
    get:
      tags: [messages]
      operationId: listMessages
      summary: List agent messages
      description: |
        Returns messages filtered by agent, status, and pagination cursor.
        Only `agent`, `status`, `limit`, and `before` are accepted query parameters —
        unknown parameters return 400 to prevent silent filtering bugs.
      parameters:
        - name: agent
          in: query
          schema:
            type: string
          description: Agent mailbox filter (returns messages TO this agent)
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, delivered, done, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 200
        - name: before
          in: query
          schema:
            type: integer
          description: Pagination cursor — message id to fetch messages before
      responses:
        "200":
          description: Message list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentMessage"
        "400":
          description: Unknown query parameter
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Error"
                  - type: object
                    properties:
                      unknown:
                        type: array
                        items:
                          type: string
                      known:
                        type: array
                        items:
                          type: string
                      hint:
                        type: string

    post:
      tags: [messages]
      operationId: sendMessage
      summary: Send an inter-agent message
      description: |
        Creates a message in the queue for delivery to the target agent's session.
        The `from` field must be a registered fleet agent id (or the owner name).
        Federation addresses use slash notation: `"peer-system/agent-name"`.

        Side effect: when `assign: true` is set and the local recipient has no `active`
        or `blocked` blackboard row, a row with `status: "assigned"` is automatically
        opened. The summary is the first 100 characters of the message content. The
        agent's own `POST /api/blackboard` with `status: "active"` overwrites this row.
        Omitting `assign` or setting it to `false` produces no blackboard side effect.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [from, to, content]
              properties:
                from:
                  type: string
                  description: Sender agent id (registered fleet agent)
                to:
                  type: string
                  description: Recipient agent id, or `"system/agent"` for federation
                content:
                  type: string
                  minLength: 1
                origin_note:
                  type: string
                  maxLength: 120
                  description: Optional attributability tag
                assign:
                  type: boolean
                  default: false
                  description: |
                    When true, opens an "assigned" blackboard row for the local recipient
                    if they have no active or blocked row. Use for deliberate task
                    delegation; omit for replies, notifications, and acknowledgements.
            example:
              from: "agent-b"
              to: "agent-a"
              content: "feat/openapi-spec kész, 23ec40d, Jónás tesztre vár"
      responses:
        "200":
          description: Message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentMessage"
        "400":
          description: |
            Missing required fields or invalid federated address.
            `error` is a stable snake_case code; human detail is in `hint`.
            Codes: `required`, `invalid_value`, `federation_disabled`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: |
            Sender not authorised.
            `error` is a stable snake_case code; human detail is in `hint`.
            Codes: `sender_reserved`, `federated_sender_not_allowed`,
            `sender_not_in_allowlist`, `unknown_sender`.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /messages/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    put:
      tags: [messages]
      operationId: updateMessageStatus
      summary: Mark a message done or failed (executor callback)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [done, failed]
                result:
                  type: string
                  description: Short result summary (first 500 chars stored)
            example:
              status: "done"
              result: "feat/openapi-spec pushed, 23ec40d"
      responses:
        "200":
          description: Status updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
        "404":
          description: Message not found or invalid status transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ---- Kanban -------------------------------------------------------------
...